+++ schema/table.go
@@ -23,9 +23,6 @@ type TableResolver func(ctx context.Context, meta ClientMeta, parent *Resource, 
 
 type RowResolver func(ctx context.Context, meta ClientMeta, resource *Resource) error
 
-// Classify error and return it's severity and type
-type IgnoreErrorFunc func(err error) (bool, string)
-
 type Tables []*Table
 
 type Table struct {
@@ -39,8 +36,6 @@ type Table struct {
 	Relations Tables `json:"relations"`
 	// Resolver is the main entry point to fetching table data and
 	Resolver TableResolver `json:"-"`
-	// IgnoreError is a function that classifies error and returns it's severity and type
-	IgnoreError IgnoreErrorFunc `json:"-"`
 	// Multiplex returns re-purposed meta clients. The sdk will execute the table with each of them
 	Multiplex func(meta ClientMeta) []ClientMeta `json:"-"`
 	// PostResourceResolver is called after all columns have been resolved, but before the Resource is sent to be inserted. The ordering of resolvers is:
@@ -49,8 +44,6 @@ type Table struct {
 	// PreResourceResolver is called before all columns are resolved but after Resource is created. The ordering of resolvers is:
 	//  (Table) Resolver → PreResourceResolver → ColumnResolvers → PostResourceResolver
 	PreResourceResolver RowResolver `json:"-"`
-	// Options allow modification of how the table is defined when created
-	Options TableCreationOptions `json:"options"`
 
 	// IgnoreInTests is used to exclude a table from integration tests.
 	// By default, integration tests fetch all resources from cloudquery's test account, and verify all tables
@@ -66,12 +59,6 @@ type Table struct {
 	Serial string `json:"-"`
 
 	columnsMap map[string]int
-}
-
-// TableCreationOptions allow modifying how table is created such as defining primary keys, indices, foreign keys and constraints.
-type TableCreationOptions struct {
-	// List of columns to set as primary keys. If this is empty, a random unique ID is generated.
-	PrimaryKeys []string
 }
 
 func (tt Tables) TableNames() []string {
@@ -169,34 +156,23 @@ func (t Table) Resolve(ctx context.Context, meta ClientMeta, syncTime time.Time,
 	go func() {
 		defer func() {
 			if err := recover(); err != nil {
+				sentry.WithScope(func(scope *sentry.Scope) {
+					scope.SetTag("table", t.Name)
+					sentry.CurrentHub().Recover(err)
+				})
 				stack := string(debug.Stack())
-				meta.Logger().Error().Interface("error", err).Str("table_name", t.Name).TimeDiff("duration", time.Now(), startTime).Str("stack", stack).Msg("table resolver finished with panic")
+				meta.Logger().Error().Interface("error", err).Str("table", t.Name).TimeDiff("duration", time.Now(), startTime).Str("stack", stack).Msg("table resolver finished with panic")
 			}
 			close(res)
 		}()
-		meta.Logger().Debug().Str("table_name", t.Name).Msg("table resolver started")
+		meta.Logger().Debug().Str("table", t.Name).Msg("table resolver started")
 		if err := t.Resolver(ctx, meta, parent, res); err != nil {
-			if t.IgnoreError != nil {
-				if ignore, errType := t.IgnoreError(err); ignore {
-					meta.Logger().Debug().Stack().Str("table_name", t.Name).TimeDiff("duration", time.Now(), startTime).Str("error_type", errType).Err(err).Msg("table resolver finished with error")
-					return
-				}
-			}
-			sentry.WithScope(func(scope *sentry.Scope) {
-				scope.SetTag("table", t.Name)
-				scope.SetLevel(sentry.LevelError)
-				sentry.CaptureMessage(err.Error())
-			})
-			meta.Logger().Error().Str("table_name", t.Name).TimeDiff("duration", time.Now(), startTime).Err(err).Msg("table resolver finished with error")
+			meta.Logger().Error().Str("table", t.Name).TimeDiff("duration", time.Now(), startTime).Err(err).Msg("table resolver finished with error")
 			return
 		}
-		meta.Logger().Debug().Str("table_name", t.Name).TimeDiff("duration", time.Now(), startTime).Msg("table resolver finished successfully")
+		meta.Logger().Debug().Str("table", t.Name).TimeDiff("duration", time.Now(), startTime).Msg("table resolver finished successfully")
 	}()
 	totalResources := 0
-	// we want to check for data integrity
-	// in the future we can do that as an optinoal feature via a flag
-	// pks := map[string]bool{}
-	// each result is an array of interface{}
 	for elem := range res {
 		objects := helpers.InterfaceSlice(elem)
 		if len(objects) == 0 {
@@ -207,25 +183,21 @@ func (t Table) Resolve(ctx context.Context, meta ClientMeta, syncTime time.Time,
 			resource := NewResourceData(&t, parent, syncTime, objects[i])
 			if t.PreResourceResolver != nil {
 				if err := t.PreResourceResolver(ctx, meta, resource); err != nil {
-					meta.Logger().Error().Str("table_name", t.Name).Err(err).Msg("pre resource resolver failed")
+					meta.Logger().Error().Str("table", t.Name).Err(err).Msg("pre resource resolver failed")
 				} else {
-					meta.Logger().Trace().Str("table_name", t.Name).Msg("pre resource resolver finished successfully")
+					meta.Logger().Trace().Str("table", t.Name).Msg("pre resource resolver finished successfully")
 				}
 			}
 			t.resolveColumns(ctx, meta, resource)
 			if t.PostResourceResolver != nil {
-				meta.Logger().Trace().Str("table_name", t.Name).Msg("post resource resolver started")
+				meta.Logger().Trace().Str("table", t.Name).Msg("post resource resolver started")
 				if err := t.PostResourceResolver(ctx, meta, resource); err != nil {
-					meta.Logger().Error().Str("table_name", t.Name).Stack().Err(err).Msg("post resource resolver finished with error")
+					meta.Logger().Error().Str("table", t.Name).Stack().Err(err).Msg("post resource resolver finished with error")
 				} else {
-					meta.Logger().Trace().Str("table_name", t.Name).Msg("post resource resolver finished successfully")
+					meta.Logger().Trace().Str("table", t.Name).Msg("post resource resolver finished successfully")
 				}
 			}
-			// if pks[resource.PrimaryKeyValue()] {
-			// 	meta.Logger().Error().Str("table_name", t.Name).Str("primary_key", resource.PrimaryKeyValue()).Msg("duplicate primary key found")
-			// } else {
-			// 	pks[resource.PrimaryKeyValue()] = true
-			// }
+
 			resolvedResources <- resource
 			for _, rel := range t.Relations {
 				totalResources += rel.Resolve(ctx, meta, syncTime, resource, resolvedResources)
@@ -238,22 +210,22 @@ func (t Table) Resolve(ctx context.Context, meta ClientMeta, syncTime time.Time,
 func (t Table) resolveColumns(ctx context.Context, meta ClientMeta, resource *Resource) {
 	for _, c := range t.Columns {
 		if c.Resolver != nil {
-			meta.Logger().Trace().Str("column_name", c.Name).Str("table_name", t.Name).Msg("column resolver custom started")
+			meta.Logger().Trace().Str("column_name", c.Name).Str("table", t.Name).Msg("column resolver custom started")
 			if err := c.Resolver(ctx, meta, resource, c); err != nil {
-				meta.Logger().Error().Str("column_name", c.Name).Str("table_name", t.Name).Err(err).Msg("column resolver finished with error")
+				meta.Logger().Error().Str("column_name", c.Name).Str("table", t.Name).Err(err).Msg("column resolver finished with error")
 			}
-			meta.Logger().Trace().Str("column_name", c.Name).Str("table_name", t.Name).Msg("column resolver finished successfully")
+			meta.Logger().Trace().Str("column_name", c.Name).Str("table", t.Name).Msg("column resolver finished successfully")
 		} else {
-			meta.Logger().Trace().Str("column_name", c.Name).Str("table_name", t.Name).Msg("column resolver default started")
+			meta.Logger().Trace().Str("column_name", c.Name).Str("table", t.Name).Msg("column resolver default started")
 			// base use case: try to get column with CamelCase name
 			v := funk.Get(resource.Item, strcase.ToCamel(c.Name), funk.WithAllowZero())
 			if v != nil {
 				if err := resource.Set(c.Name, v); err != nil {
-					meta.Logger().Error().Str("column_name", c.Name).Str("table_name", t.Name).Err(err).Msg("column resolver default finished with error")
+					meta.Logger().Error().Str("column_name", c.Name).Str("table", t.Name).Err(err).Msg("column resolver default finished with error")
 				}
-				meta.Logger().Trace().Str("column_name", c.Name).Str("table_name", t.Name).Msg("column resolver default finished successfully")
+				meta.Logger().Trace().Str("column_name", c.Name).Str("table", t.Name).Msg("column resolver default finished successfully")
 			} else {
-				meta.Logger().Trace().Str("column_name", c.Name).Str("table_name", t.Name).Msg("column resolver default finished successfully with nil")
+				meta.Logger().Trace().Str("column_name", c.Name).Str("table", t.Name).Msg("column resolver default finished successfully with nil")
 			}
 		}
 	}